Change field is_admin to tinyint.

+++ db/LiveSeederUser.php
@@ -25,7 +25,7 @@ class LiveSeederUser extends Model
         $sql = "CREATE TABLE IF NOT EXISTS users(
             id INT AUTO_INCREMENT PRIMARY KEY,
             gid INT NOT NULL DEFAULT 0,
-            is_admin INT NOT NULL DEFAULT 0,
+            is_admin TINYINT(1) NOT NULL DEFAULT 0,
             first_name VARCHAR(50) NOT NULL,
             last_name VARCHAR(50) NOT NULL,
             photo_url TINYTEXT,
